Properly support extracting fields from entities into projections
Fixes: #31774
