+++ integration-tests/spring-data-jpa/src/main/java/io/quarkus/it/spring/data/jpa/Post.java
@@ -1,6 +1,5 @@
 package io.quarkus.it.spring.data.jpa;
 
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -18,7 +17,7 @@ import jakarta.persistence.Table;
 
 @Entity(name = "Post")
 @Table(name = "post")
-public class Post implements ByPassHolder {
+public class Post extends AbstractPost implements ByPassHolder {
 
     @Id
     @SequenceGenerator(name = "postSeqGen", sequenceName = "postSeq", initialValue = 100, allocationSize = 1)
@@ -34,8 +33,6 @@ public class Post implements ByPassHolder {
     private Map<String, String> metadata = new HashMap<>();
 
     private boolean bypass;
-
-    private ZonedDateTime posted;
 
     private String organization;
 
@@ -69,14 +66,6 @@ public class Post implements ByPassHolder {
 
     public void setBypass(boolean bypass) {
         this.bypass = bypass;
-    }
-
-    public ZonedDateTime getPosted() {
-        return posted;
-    }
-
-    public void setPosted(ZonedDateTime postedAt) {
-        this.posted = postedAt;
     }
 
     public String getOrganization() {
